feat: change pgindent version and perltidy rc
The pg_bsd_indent version is 2.1.2, which is being used by latest
PostgreSQL release 16 and master 17. The pg_bsd_indent options used by
pgindent are kept the same as PostgreSQL 16 and master.

The perltidy version is v20230309, which is also being used by latest
PostgreSQL release and master. The perltidyrc configuration is set to
the same as PostgreSQL 16 and master, generating some diff for 15, but
we choose to follow up the latest rules.

Let the hacking of PolarDB-PG 15 begin :)